+++ substratevm/src/com.oracle.graal.pointsto/src/com/oracle/graal/pointsto/meta/AnalysisField.java
@@ -544,16 +544,21 @@ public abstract class AnalysisField extends AnalysisElement implements WrappedJa
      */
     public void beforeFieldValueAccess() {
         declaringClass.registerAsReachable(this);
-        declaringClass.ensureOnTypeReachableTaskDone();
 
-        List<AnalysisFuture<Void>> notifications = declaringClass.scheduledTypeReachableNotifications;
-        if (notifications != null) {
-            for (var notification : notifications) {
-                notification.ensureDone();
+        declaringClass.forAllSuperTypes(type -> {
+            type.ensureOnTypeReachableTaskDone();
+
+            List<AnalysisFuture<Void>> notifications = type.scheduledTypeReachableNotifications;
+            if (notifications != null) {
+                for (var notification : notifications) {
+                    notification.ensureDone();
+                }
+                /*
+                 * Now we know all the handlers have been executed, no checks are necessary anymore.
+                 */
+                type.scheduledTypeReachableNotifications = null;
             }
-            /* Now we know all the handlers have been executed, no checks are necessary anymore. */
-            declaringClass.scheduledTypeReachableNotifications = null;
-        }
+        });
     }
 
     public void addAnalysisFieldObserver(AnalysisFieldObserver observer) {