+++ substratevm/src/com.oracle.svm.hosted/src/com/oracle/svm/hosted/snippets/ReflectionPlugins.java
@@ -635,10 +635,19 @@ public final class ReflectionPlugins {
 
         /* Any other object that is not a Class. */
         Object result = snippetReflection.asObject(Object.class, argConstant);
-        if (result != null && ALLOWED_CONSTANT_CLASSES.contains(result.getClass())) {
+        if (result != null && isAllowedConstant(result.getClass())) {
             return result;
         }
         return null;
+    }
+
+    private boolean isAllowedConstant(Class<?> clazz) {
+        for (var allowed : ALLOWED_CONSTANT_CLASSES) {
+            if (allowed.isAssignableFrom(clazz)) {
+                return true;
+            }
+        }
+        return false;
     }
 
     /**