+++ vm/mx.vm/suite.py
@@ -1,6 +1,6 @@
 suite = {
     "name": "vm",
-    "version" : "24.0.0",
+    "version" : "24.1.0",
     "mxversion": "6.41.0",
     "release" : False,
     "groupId" : "org.graalvm",
@@ -57,7 +57,7 @@ suite = {
             },
             {
                 "name": "fastr",
-                "version": "e67d35460eb33d47cfe1a2d535b64524eacfed81",
+                "version": "ea083f09c707b48b25d58fda6cb290187726d9f6",
                 "dynamic": True,
                 "urls": [
                     {"url": "https://github.com/oracle/fastr.git", "kind": "git"},
@@ -65,7 +65,7 @@ suite = {
             },
             {
                 "name": "graalpython",
-                "version": "cf3c107181f2375f4241eba4bc49751b43f5bd88",
+                "version": "10ee90e185b0ecdc5ad6b7313b146faaf8dc51f9",
                 "dynamic": True,
                 "urls": [
                     {"url": "https://github.com/graalvm/graalpython.git", "kind": "git"},